requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-polymorphic>=3.1.0
psycopg2-binary>=2.9.5
python-decouple>=3.8
django-cors-headers>=4.3.1
celery>=5.3.0
redis>=5.0.0
Pillow>=10.0.0
django-storages>=1.14.0
boto3>=1.34.0
stripe>=7.0.0
django-allauth>=0.57.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
django-extensions>=3.2.3
django-debug-toolbar>=4.2.0
django-environ>=0.11.0
gunicorn>=21.2.0
whitenoise>=6.6.0
cryptography>=42.0.0
django-filter>=24.2
scikit-learn>=1.3.0
numpy>=1.24.0
